Added Save Queries as Rules

+++ scripts/kibanaBuild.sh
@@ -8,12 +8,37 @@ set -x
 # will NOT build kibana from scratch, instead it will use that .tar.gz file to create
 # the rpm
 
+if [[ $# -gt 2 || $# -lt 1 ]] ; then
+    echo 'Usage:  sh kibanaBuild.sh <PROTOBUFFER-BRANCH><PROTOBUFFER-GIT-USER<optional>>'
+    exit 0
+fi
+
 rm -rf ../kibana-build-target
 mkdir -p ../kibana-build-target
 cp -r . ../kibana-build-target/.
 LOCATION=$PWD
 cd ../kibana-build-target
 
+BRANCH="$1"
+
+if [ $# -eq 2 ]; then
+   USER=$2
+else
+   USER="Logrhythm"
+fi
+echo "USER IS $USER";
+
+echo "Building: $BRANCH for USER: $USER"
+
+git clone git@lrgit:$USER/Protobuffers.git -b $BRANCH
+
+kibanaBuildDir=$PWD
+
+cd Protobuffers
+sh scripts/buildUIFieldMap.sh
+cp js/fieldMap.js $kibanaBuildDir/src/kibana/netmon_libs/
+cd $kibanaBuildDir
+rm -rf Protobuffers
 
 sed -i s/\'shasum/\'sha1sum/g tasks/create_shasums.js
 sed -i s/0.10.x/0.10.42/g .node-version
